refactor(oxlint/linter/lsp): move ignore logic to `oxc_linter` (#24951)
The start for fixing https://github.com/oxc-project/oxc/issues/24570

At the moment the "ignore this line" / "ignore this section" logic is living inside `oxlint/lsp` directory.
Because we now need to know the AST parents, the code should live inside `oxc_linter`.

This PR refactors the 2 methods into a new `oxc_linter` file.
